+++ src/main/java/chess/api/storage/ephemeral/InMemoryTrie.java
@@ -6,33 +6,33 @@ public class InMemoryTrie {
 
     private static final IntArrayComparator SHORT_ARRAY_COMPARATOR = new IntArrayComparator();
 
-    private final Map<int[], int[]> trieMap = new TreeMap<>(SHORT_ARRAY_COMPARATOR);
+    private final Map<short[], short[]> trieMap = new TreeMap<>(SHORT_ARRAY_COMPARATOR);
 
     public InMemoryTrie() {
-        trieMap.put(new int[]{}, null);
+        trieMap.put(new short[]{}, null);
     }
 
-    public Optional<int[]> getAvailableMoves(int[] movesSoFar) {
-        final Optional<int[]> optionalTrieNode = getNodeAtPath(movesSoFar);
+    public Optional<short[]> getAvailableMoves(short[] movesSoFar) {
+        final Optional<short[]> optionalTrieNode = getNodeAtPath(movesSoFar);
         if (optionalTrieNode.isPresent()) {
-            final int[] onwardMoves = optionalTrieNode.get();
+            final short[] onwardMoves = optionalTrieNode.get();
             return Optional.of(onwardMoves);
         }
         return Optional.empty();
     }
 
-    public void setAvailableMoves(int[] movesSoFar, int[] availableMoves) {
+    public void setAvailableMoves(short[] movesSoFar, short[] availableMoves) {
         trieMap.put(movesSoFar, availableMoves);
     }
 
-    public Optional<int[]> getNodeAtPath(int[] movesSoFar) {
+    public Optional<short[]> getNodeAtPath(short[] movesSoFar) {
         if (movesSoFar == null) {
             return Optional.empty();
         }
         return Optional.ofNullable(trieMap.get(movesSoFar));
     }
 
-    public void prune(int[] branchToPreserve) {
+    public void prune(short[] branchToPreserve) {
         trieMap.keySet().removeIf(
             moveHistory -> {
                 int comparison = SHORT_ARRAY_COMPARATOR.compare(moveHistory, branchToPreserve);
